feat: Max fee (#66)
* feat: Max fee validation

* test: Revert if max fee bp is violated (atomic)

* test: Partial success if max fee bp is violated (non-atomic)

* refactor: Extract common code

* refactor: Extract common code in multiple markets test

* perf: Re-use already loaded fee bp

* test: Set fee tests

* perf: Make bp uint256

* fix: Variable shadowing warning

+++ contracts/libraries/OrderStructs.sol
@@ -22,6 +22,6 @@ struct TokenTransfer {
 }
 
 struct FeeData {
-    uint16 bp;
+    uint256 bp;
     address recipient;
 }
